requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-ckeditor==6.2.0
django-cleanup==5.2.0
django-cors-headers==3.10.1
django-js-asset==2.0.0
djangorestframework==3.13.1
passlib==1.7.4
Pillow==9.0.1
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
